requirements/base.txt
# Frameworks
Django>=1.9,<1.10.0

# Connectors
psycopg2>=2.6.0,<2.7.0

# Plugins
django-redis-cache>=1.6.5,<1.7.0
sqlalchemy>=1.0.13,<1.1.0
django_compressor>=2.0,<2.1
requests>=2.10.0,<2.11.0

# Celery
celery>=3.1.23,<3.2.0
msgpack-python==0.4.6,<0.5.0

# Django RESTful Framework
djangorestframework>=3.3.3,<3.4.0
django-filter>=0.13.0,<0.14.0
markdown>=2.6.6,<2.7.0
django-cors-headers==1.1.0,<1.2.0

# OAuth and JWT
oauthlib>=1.1.2,<1.2.0
python-jose==1.0.0,<1.1.0

